automation_playwright/requirements.txt
playwright
pytest-playwright
